+++ betarun/GC/objectsize.c
@@ -1,6 +1,6 @@
 /*
  * BETA RUNTIME SYSTEM, Copyright (C) 1991 Mjolner Informatics Aps.
- * Mod: $RCSfile: objectsize.c,v $, rel: %R%, date: $Date: 1992-06-08 15:17:24 $, SID: $Revision: 1.7 $
+ * Mod: $RCSfile: objectsize.c,v $, rel: %R%, date: $Date: 1992-06-10 14:50:59 $, SID: $Revision: 1.8 $
  * by Lars Bak
  */
 #include "beta.h"
@@ -12,21 +12,21 @@ long ObjectSize(theObj)
   ref(ProtoType) theProto = theObj->Proto;
 
   if( (long) theProto < 0 ){
-    switch((long) theProto){
-    case ByteRepPTValue:
+    switch((int) theProto){
+    case (int) ByteRepPTValue:
       /* BodySize= ((Range+1+3)/4): One byte extra for \0 in strings and multiple of 4 */
       return (headsize(ValRep) + (toValRep(theObj)->HighBorder+4)) >> 2;
-    case ValRepPTValue:
+    case (int) ValRepPTValue:
       return (headsize(ValRep)>>2) + toValRep(theObj)->HighBorder;
-    case RefRepPTValue:
+    case (int) RefRepPTValue:
       return (headsize(RefRep)>>2) + toRefRep(theObj)->HighBorder;
-    case ComponentPTValue:
+    case (int) ComponentPTValue:
       return (headsize(Component)>>2)
 	+ (ComponentItem(theObj)->Proto->Size + 1);
-    case StackObjectPTValue:
+    case (int) StackObjectPTValue:
       return (headsize(StackObject)>>2)
 	+ toStackObject(theObj)->ObjectSize;
-    case StructurePTValue:
+    case (int) StructurePTValue:
       return (headsize(Structure)>>2);
     }
   }else
